API Issues and download master changes

+++ src/main/java/com/iemr/mmu/repo/nurse/covid19/Covid19BenFeedbackRepo.java
@@ -21,12 +21,14 @@
 */
 package com.iemr.mmu.repo.nurse.covid19;
 
+import java.math.BigInteger;
+
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 import com.iemr.mmu.data.covid19.Covid19BenFeedback;
 
 @Repository
-public interface Covid19BenFeedbackRepo extends CrudRepository<Covid19BenFeedback, Long> {
+public interface Covid19BenFeedbackRepo extends CrudRepository<Covid19BenFeedback, BigInteger> {
 	Covid19BenFeedback findByBeneficiaryRegIDAndVisitCode(Long benRegID, Long visitCode);
 }
